Check for SGX and necessary compiler flags during configuration.
Use -fstack-protector-strong instead of -fstack-protector when available.

Move the SGX checks
Check if SGX is enabled before testing for any headers (like pthread).

Check the right headers when in SGX mode.
Neither signal.h nor sys/mman.h are present in the SGX version of libc.

Add some SGX-specific code for thread-safety.
Use SGX multi-threading code for the critical sections guards.

Fix the RNG for SGX
Use SGX API (sgx_read_rand) to generate random bytes.
We still need to figure something out for the SGX simulation mode (the current code is insecure in this mode).

Remove some function checks for SGX builds.
We know some of them won't be available, and for a reason I do not understand, they still might me detected by autoconf.

Fix the Salsa20 RNG for SGX.
Use sgx_read_rand to see the stream cipher generating the random bytes.

Embed bitcode in iOS builds

Disastrous typo

Beginning to develop SGX unit tests

Do not put .so files in the source tree.

Improve the autoconf and automate files.
What we modify with SGX when compiling the library is the include path, so let's restrict ourselves to only change CPPFLAGS during the configuration.
A specific SGX_CPPFLAGS variable is created and used during the compilation of the library. Note that it is empty by default.

A first functional test (aead_aes256_gcm).
The other tests should be easy to construct from this base.

Unifying the test Makefile
Try to have a Makefile that differs only a little between the SGX and non-SGX mode.

Fix an issue with the include path.
The code determining wether AES-NI is supported or not was bugged.

Tests compile for both the SGX and regular target

Correctly build test apps for SGX

Add explicit rules to build the enclaves.
It happens that make rules with pattern do not work with automake (they are not portable).
Having explicit rules seems to work, but is ugly. It would be cool to automatically generate them using autoconf.

All tests compile, many (but not all) pass.

Fix the rand() function for SGX

Fix the sodium_core and sodium_utils{2,3} test.
All the tests now pass on an SGX build. Though it is not clear that sodium_utils3 actually tests anything on SGX.

Fix autoconf

Fix the rand() placeholder for SGX
rand() has to return a non negative value.

Increase the maximum heap size of the test enclaves.
Test programs used to allocate more memory than allowed by the enclave, and thus miserably failed.

Remove obsolete setting FUNCTION_POINTER_ALIGNMENT

FUNCTION_POINTER_ALIGNMENT was removes from emscripten in January 2018:
https://github.com/kripken/emscripten/pull/6091

Clean up the automake file
Use a suffix rule to compile the enclaves.
Automatically generate the enclaves filenames

Distribute the new files required by an SGX build

Properly clean enclaves

Fixed build command pretty-printing

Remove commented code

Link against the right SGX runtime libraries when in hardware mode

Add a configure option and fix the automake file
Automate does not support ifeq statements: they are not copied to the the generated makefile(s). Instead, use regular if statement with conditions defined in configure.ac.
As a consequence, add new a configuration option build the test programs in hardware mode when SGX is enabled.

Fix the distcheck target
enclave_u.h and enclave_u.c are intermediates that should
not be included in the distribution.
Also, because I fixed a typo in the test/default/Makefile.am file
(*_SOURCE -> *_SOURCES for the test targets), I had to define quirks.h source
dependencies as nodist_EXTRA. Note that quirks/quirks.h is already included in
the distribution by test/Makefile.am, so this is not a problem. Yet, it could
simplify things if we put quirks.h in the same directory as the other files
used by the test.

Stupid typo...
It is sgx_hw_tests (with an 's') and not sgx_hw_test (without the 's').

Change the enclave sources from the nodist_EXTRA_ to the nodist_ section.
Using EXTRA apparently prevents make from compiling the intermediates.

Make test_app.c explicitely depend on enclave_u.h
Otherwise, make might not create enclave_u.h on time,
before compiling test_app.c, causing a compilation error.

Typo

Add logging to enclave tests.
This should allow for better debugging of the enclave tests.

Immediately flush the log messages to sdtout.
Just in case we have an error killing the program before the
sdout buffer is flushed ...

Replace sodium_malloc by a memory pool for enclave tests.
sodium_malloc (and most of the memory management functions of utils.h)
use syscalls, are not supported inside an enclave (it raises a
SIGILL), which means that, to run the tests inside an enclave, we must allocate
the memory otherwise. I just reused the code of cmptest.h in benchmark
mode.
Also, the sodium_utils3 test is now disabled during an SGX build.